+++ pyjoulescope_driver/entry_points/api_timeout.py
@@ -33,38 +33,27 @@ def parser_config(p):
     p.add_argument('--threads',
                    type=int,
                    default=10,
-                   help='The number of threads to run simultaneously for testing.')
+                   help='The number of threads to run simultaneously for API timeout testing.')
     p.add_argument('--timeout', '-t',
                    type=float,
-                   default=0.01,
-                   help='The thread processing timeout delay in seconds.')
+                   default=0.25,
+                   help='The API timeout in seconds.')
     return on_cmd
-
-
-def _device_get(d):
-    devices_paths = d.device_paths()
-    if len(devices_paths) == 1:
-        return devices_paths[0]
-    if len(devices_paths) == 0:
-        print('No devices found')
-        return None
-    elif len(devices_paths) > 1:
-        device_path = devices_paths[0]
-        print(f'Multiple devices found, selecting {device_path}')
-        return device_path
 
 
 def _thread_run(state):
     global _quit
     index = state['index']
     d = state['driver']
-    device_path = state['device_path']
-    timeout_ms = state['timeout_ms']
-    timeout_api = state['timeout_api']
     counter = 0
     while not _quit:
-        v64 = timeout_ms | (index << 56) | (counter << 32)
-        d.publish(device_path + '/h/timeout', v64, timeout=timeout_api)
+        try:
+            d.publish('@/timeout', 0, timeout=state['timeout'])
+            if state['timeout']:
+                print('No timeout when expected')
+                return 1
+        except TimeoutError:
+            pass  # expected
         counter += 1
     state['counter'] = counter
     print(f'{index} {counter}')
@@ -74,26 +63,13 @@ def on_cmd(args):
     global _quit
     with Driver() as d:
         d.log_level = args.jsdrv_log_level
-        device_path = _device_get(d)
-        if device_path is None:
-            return 1
-        d.open(device_path)
-        try:
-            d.publish(device_path + '/h/timeout', 100, timeout=0.001)
-            print('No timeout when expected')
-            return 1
-        except TimeoutError:
-            pass  # expected
-
         threads = []
         for idx in range(args.threads):
             state = {
                 'name': f'jsdrv {idx}',
                 'index': idx,
                 'driver': d,
-                'device_path': device_path,
-                'timeout_ms': int(1000 * args.timeout),
-                'timeout_api': args.timeout * args.threads * 2 + 0.25,
+                'timeout': args.timeout,
                 'counter': 0,
             }
             state['thread'] = threading.Thread(name=state['name'], target=_thread_run, args=[state])
@@ -110,6 +86,5 @@ def on_cmd(args):
             thread['thread'].join()
             counter += thread['counter']
         print(f'Total calls: {counter}')
-
 
     return 0